Initial design of ingest infrastructure.
Speeds are very meh. The bottleneck is either InfluxDB or the official
Java-library that I use to communicate with it (which communites with the
database through a REST-api, so there is lots of overhead)

Ingest speed on my windows machine: ~35000 entries per second.
Ingest speed on the provided linux server: ~21000 entries per second.

Ingest speed on my windows machine when the final insert-step to the db
for generated entries is disabled: ~4 mill per second.

All of this is with 1 ingest-thread and the database running on the same
machine as the benchmark.