Change group-by interval of 'floor-totals' and 'total-clients'
Changed the interval from however many seconds were between readings (to
match the pre-computations done by ITU) to grouping by day instead.
The motivation for this change is to reduce the significant overhead of
transfering data from the database to the benchmark-program for every AP
in every group.

This obviously also decreases the number of groups and therefore the
amount of work that the database needs to do, but... eh.