Add support for column-schema format.
This commit adds support for changing the schema used by the benchmark
during generation and querying.

Alongside this change, ITarget- and IQueries-implementations for both
Influx and Timescale have been written and the existing database-code
was heavily refactored.

The python script to compare query-results was also updated and can
check comparability of query-results across schema-format despite the
minor changes in query-results (due to how single missing AP values have
to be represented).