Add support for Apache Kudu
This adds preliminary support for Apache Kudu. Data can be inserted into
Kudu in both ROW and COLUMN format, and it can be queried in both
formats as well.
Query-results have been compared against those returned by both
Timescale and Influx in both formats and are identical when compared
with 'comparequeries.py'.

This initial implementation has only been tested against a known, static
dataset. Changes may need to be done to get satisfactory performance on
other datasets. See TODOs in comments.

Mainly, these issues are related to partitioning of data and to the
in-application implementations of typical SQL-primitives that are
missing in the Kudu-libraries severely limited API.