Fix duplicate key-error for Kudu row-format.
With e.g. a second-granularity, row-format will generate the same
timestamp for most of the sensor-inserts. This isn't an issue in
'normal' time-series systems since they handle the time-column in a
special way. However, Kudu 1) requires a key on tables and 2) requires
that the key is unique. So we are forced to declare a composite-key of
(time,AP).

That choice of composite-key will be fine unless the configured
granularity is coarser than the desired sample rate, in which case this
issue will appear again. However, noone should realistically want that
coarse a sampling rate...