Fix race between ingestion and querying on empty database
This attempts to fix the race-condition present when both ingestion and
querying are configured to run against an (initially) empty table by
delaying querying until the table has a valid value.

Additionally, it catches the invalid configuration where the earliest
query date is set later than the timestamps present in the database.

However, some weird (but potentially valid) configurations may hit these
edge-cases so we warn and attempt to fix them if possible, and if not we
simply continue on and rely on asserts to trip if the config is actually
invalid.