Add K-Means query.
K-Means has been setup in the query-runner, config-settings were added
and a class that does the computation now exists. Each
database-implementation just needs to pass a function that pulls data
from the database and parses it into a time-series format for the
computation to use.

Implemented for influx w/column config for now.