Implement K-Means for all database/schema combos
This also updates the comparequeries.py script to be able to verify that
the results are identical regardless of database-choice.

This implementation has the caveat that K-Means query-results aren't
100% identical between the ROW- and COLUMN-schema due to 2 factors:
1) There are differences in how missing values are represented. For the
   row-schema missing values are just absent, whereas for the
   column-schema missing values have a value of 0 since not all
   databases support NULL-values.
2) The naive K-Means distance function doesn't take timestamps into
   account.

This means that for a column-schema an AP might have these client-values:
3 3 3 3 6 8 9 9 0 0 0 0 0 0 0 0 3 6 4 2 1
whereas for the row-schema the same AP has these client-values:
3 3 3 3 6 8 9 9                 3 6 4 2 1
But since we dont look at the timestamp in the data for the K-Means
distance function, we cannot tell that there is a hole and we therefore
end up comparing offset data with non-offset data when not all APs have
the same holes.

However, this isn't an issue that affects query-performance as reported
by the benchmark because we do not do early termination once converged.
That is, we always execute the same number of queries against the
database despite this issue.