Redact hardcoded ITU seed metadata.
The git-history will be rewritten to exclude any old versions of these
files to ensure that the ITU seed metadata isn't present.

While this probably isn't necessary because:
1) this metadata is just a mapping between
   'access point name' <-> 'floor' <-> 'location'
   with some additional info about data-weirdness
2) and the naming-schema changed wildly just after I began working with
   this to simplify the system, so this metadata is already outdated
   anyway.
However, I'm redacting it because it's easier than asking for
permission to share it, and I'd need to rewrite the seed-loading
component anyway to make the benchmark useful for non-ITU seed data
use-cases.

The version of the git repository prior to history being rewritten
(which is the benchmark-version used during my thesis) can be shared
upon request by people from the ITU (just send me an email from your
@itu.dk email address).