Add a simple seed data generator.
This generator is NOT intended to create realistic datasets, but rather
as a data-set for those who do not have their own data to use, or as a
starting point for people who want to evaluate the benchmark before
exporting their own data.

Additionally, it helps document the expected format of the seed files.

This should not be used for any actual benchmarking because the cache-
and compression-behaviors of databases using data generated by this
program are likely to be very different from real data.